Day06/Legal & Compliance Risk Identifier/requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.1
langchain==0.1.9
langchain-google-genai==0.0.11
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.2 